changed from long to int

+++ stack/stack.h
@@ -2,14 +2,14 @@
 
 typedef struct s_stack
 {
-	long			n;
+	int			n;
 	struct s_stack	*next;
 }					t_stack;
 
 t_stack				*ft_stnew(long n);
 void				ft_stpush(t_stack **stack, t_stack *content);
 void				ft_stpop(t_stack **stack);
-long				ft_stpeek(t_stack *stack);
+int				ft_stpeek(t_stack *stack);
 int					ft_stisempty(t_stack *stack);
 int					ft_stsize(t_stack *stack);
 void				ft_stclear(t_stack **stack);
